app/requirements.txt
Flask>=2.1
flask-cors
python-dotenv
gunicorn

